moved logging to general package
