+++ src/sushi_train/general/__init__.py
@@ -1,0 +1,1 @@
+from .logging import rotating_logger_json
